cleaned comments + added new file
